Extract Stopwatch and TimeSpan from TestPerformance
